api/assets/requirements.txt
flask==2.0.1
google-cloud-firestore==2.3.0
flask_cors==3.0.10
